Refactor "spaghetti" code
