+++ src/androidTest/java/medic/gateway/alert/test/InstrumentationTestUtils.java
@@ -2,11 +2,29 @@ package medic.gateway.alert.test;
 
 import android.content.Context;
 import android.content.SharedPreferences;
+import android.view.View;
+import android.widget.TextView;
+import androidx.lifecycle.Lifecycle;
+import androidx.test.core.app.ActivityScenario;
+import androidx.test.espresso.NoMatchingViewException;
+import androidx.test.espresso.PerformException;
+import androidx.test.espresso.ViewAssertion;
 import androidx.test.ext.junit.rules.ActivityScenarioRule;
 import java.util.Iterator;
 import medic.gateway.alert.*;
 import org.json.*;
+import static androidx.test.core.app.ApplicationProvider.getApplicationContext;
+import static androidx.test.espresso.Espresso.onView;
+import static androidx.test.espresso.action.ViewActions.closeSoftKeyboard;
+import static androidx.test.espresso.action.ViewActions.scrollTo;
+import static androidx.test.espresso.action.ViewActions.typeText;
+import static androidx.test.espresso.assertion.ViewAssertions.matches;
+import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
+import static androidx.test.espresso.matcher.ViewMatchers.isNotChecked;
+import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.platform.app.InstrumentationRegistry.*;
+import static medic.gateway.alert.R.*;
+import static medic.gateway.alert.test.BuildConfig.IS_MEDIC_FLAVOUR;
 import static org.junit.Assert.*;
 
 public final class InstrumentationTestUtils {
@@ -90,5 +108,64 @@ public final class InstrumentationTestUtils {
 			return true;
 		}
 		return false;
+	}
+
+	public static void assertNotChecked(int cbxId) {
+		onView(withId(cbxId)).check(matches(isNotChecked()));
+	}
+
+	public static void assertErrorDisplayed(int errorMessageResourceId) {
+		int componentId = IS_MEDIC_FLAVOUR ? id.txtWebappInstanceName : id.txtWebappUrl;
+		assertErrorDisplayed(componentId, errorMessageResourceId);
+	}
+
+	public static void assertErrorDisplayed(int componentId, int errorMessageResourceId) {
+		String errorString = getApplicationContext().getResources().getString(errorMessageResourceId);
+		assertErrorDisplayed(componentId, errorString);
+	}
+
+	public static void assertErrorDisplayed(int componentId, final String expectedMessage) {
+		onView(withId(componentId))
+				.check(new ViewAssertion() {
+					public void check(View view, NoMatchingViewException noViewFoundException) {
+						if(!(view instanceof TextView))
+							fail("Supplied view is not a TextView, so does not have an error property.");
+						TextView tv = (TextView) view;
+						assertEquals(expectedMessage, tv.getError());
+					}
+				});
+	}
+
+	public static void assertVisible(int viewId) {
+		try {
+			onView(withId(viewId)).perform(scrollTo()).check(matches(isDisplayed()));
+		} catch (PerformException e) {
+			// The view is fully visible, and the element
+			// should be displayed without the need to scroll down
+			onView(withId(viewId)).check(matches(isDisplayed()));
+		}
+	}
+
+	@SuppressWarnings("PMD.EmptyCatchBlock")
+	public static void assertDoesNotExist(int viewId) {
+		try {
+			onView(withId(viewId)).check(matches(isDisplayed()));
+			fail("Found view which should not exist!");
+		} catch(NoMatchingViewException ex) {
+			// expected
+		}
+	}
+
+	public static Settings settings() {
+		return Settings.in(getApplicationContext());
+	}
+
+	public static SettingsStore settingsStore() {
+		return SettingsStore.in(getApplicationContext());
+	}
+
+	public static void enterText(int componentId, String text) {
+		onView(withId(componentId))
+				.perform(typeText(text), closeSoftKeyboard());
 	}
 }
